requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.29.7
twilio==8.10.0
requests==2.31.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

